requirements.txt
playwright==1.48.0
google-generativeai==0.8.0
python-dotenv==1.0.0
Pillow==10.4.0
aiofiles==24.1.0
